+++ openbas-api/src/main/java/io/openbas/utils/AtomicTestingUtils.java
@@ -1,9 +1,9 @@
 package io.openbas.utils;
 
+import io.openbas.atomic_testing.TargetType;
 import io.openbas.database.model.*;
 import io.openbas.database.model.InjectExpectation.EXPECTATION_TYPE;
 import io.openbas.database.raw.*;
-import io.openbas.atomic_testing.TargetType;
 import io.openbas.expectation.ExpectationType;
 import io.openbas.rest.atomic_testing.form.InjectTargetWithResult;
 import io.openbas.utils.InjectMapper.ExpectationResultsByType;
